+++ tests/conftest.py
@@ -1,5 +1,5 @@
 # pylint: disable=redefined-outer-name
-from os import chdir, getcwd, listdir, path
+from os import chdir, getcwd, path
 from subprocess import check_output
 from typing import List
 
@@ -9,10 +9,9 @@ import pytest
 ACTUAL_CWD = getcwd()
 
 
-def protostar(args: List[str]) -> str:
-    return check_output(
-        ["python", path.join(ACTUAL_CWD, "protostar.py")] + args
-    ).decode("utf-8")
+@pytest.fixture(autouse=True)
+def change_cwd(tmpdir):
+    return chdir(tmpdir)
 
 
 @pytest.fixture
@@ -20,18 +19,7 @@ def project_name():
     return "foobar"
 
 
-@pytest.fixture(autouse=True)
-def change_cwd(tmpdir):
-    return chdir(tmpdir)
-
-
-def test_help():
-    result = protostar(["--help"])
-
-    assert "usage:" in result
-
-
-def test_init(project_name: str):
+def init_project(project_name: str):
     child = pexpect.spawn(f"python {path.join(ACTUAL_CWD, 'protostar.py')} init")
     child.expect("Project name:", timeout=5)
     child.sendline(project_name)
@@ -46,4 +34,13 @@ def test_init(project_name: str):
     child.expect("Libraries directory *", timeout=1)
     child.sendline("")
 
-    assert "package.toml" in listdir(f"./{project_name}")
+
+def protostar(args: List[str]) -> str:
+    return check_output(
+        ["python", path.join(ACTUAL_CWD, "protostar.py")] + args
+    ).decode("utf-8")
+
+
+@pytest.fixture
+def init(project_name: str):
+    init_project(project_name)
